+++ client/src/components/layouts/MismatchOverlay/index.ts
@@ -1,0 +1,2 @@
+export { default as BasicMismatchOverlay } from './BasicMismatchOverlay';
+export * from './BasicMismatchOverlay';